test(models): verify generated invocation contracts
